requirements.txt
streamlit>=1.30.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
faiss-cpu>=1.7.4
chromadb>=0.4.22
neo4j>=5.15.0
pypdf>=4.0.0
pdfminer.six>=20231228
tiktoken>=0.5.2
pytest>=8.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
altair>=5.0.0
langchain-chroma>=0.1.0
